fix marked points
