+++ product/src/main/java/com/epam/esm/service/OrderService.java
@@ -15,5 +15,4 @@ public interface OrderService {
 
     List<OrderResponseDTO> findUserOrders(Long userId);
 
-    boolean isUser(String token);
 }
