+++ product/src/main/resources/locale/locale_en.properties
@@ -15,4 +15,6 @@ locale.message.InvalidToken=JWT token is expired or invalid
 locale.message.error.TokenNotValid=Token is not valid, try to login again
 locale.message.UserAlreadyExist=Such user has already created
 locale.message.error.EntityAlreadyExist=Entity has already created
+locale.message.error.SomethingWentWrong=Something went wrong, sorry it is our mistake.\
+We are repearing
