+++ product/src/main/resources/locale/locale_ru.properties
@@ -15,3 +15,5 @@ locale.message.InvalidToken=Неверный или просроченный"Токен"
 locale.message.error.TokenNotValid=Токен не валидный, попробуйте залогиниться снова
 locale.message.UserAlreadyExist=Такой пользователь уже создан
 locale.message.error.EntityAlreadyExist=Сущьность уже создана
+locale.message.error.SomethingWentWrong=Что-то пошло не так, простите это наша ошибка.\
+Мы ремонтируем
